refactored wallet to ddd

+++ migrations/20250823222415_add_wallet_table.sql
@@ -2,8 +2,8 @@
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS wallets (
   user_id BIGINT PRIMARY KEY REFERENCES users(id),
-  balance NUMERIC(12,2) NOT NULL DEFAULT 0,
-  reserved NUMERIC(12,2) NOT NULL DEFAULT 0
+  balance INT NOT NULL DEFAULT 0,
+  reserved INT NOT NULL DEFAULT 0
 );
 -- +goose StatementEnd
